+++ src/renderer/const.ts
@@ -10,7 +10,6 @@ import { GAIAChain } from '@xchainjs/xchain-cosmos'
 import { DASHChain } from '@xchainjs/xchain-dash'
 import { DOGEChain } from '@xchainjs/xchain-doge'
 import { ETHChain } from '@xchainjs/xchain-ethereum'
-import { KUJIChain } from '@xchainjs/xchain-kujira'
 import { LTCChain } from '@xchainjs/xchain-litecoin'
 import { MAYAChain } from '@xchainjs/xchain-mayachain'
 import { RadixChain } from '@xchainjs/xchain-radix'
@@ -227,7 +226,6 @@ export const CHAIN_WEIGHTS_THOR: Record<EnabledChain, number> = {
   [GAIAChain]: 10,
   [ARBChain]: 11,
   [DASHChain]: 12,
-  [KUJIChain]: 13,
   [RadixChain]: 14,
   [BASEChain]: 15,
   [ADAChain]: 16,